Move the encoder logic out of lib.rs
